% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/determine_letter_indx.R
\name{determine_letter_indx}
\alias{determine_letter_indx}
\title{Determine letter indexes}
\usage{
determine_letter_indx(char)
}
\arguments{
\item{char}{A character vector of length one}
}
\description{
This function takes a charactor vector of length one and
returns a character vector of numbers corresponding
with letter indexes in the letters object.
}

