% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hpybdy.R
\docType{package}
\name{hpybdy}
\alias{hpybdy}
\alias{hpybdy-package}
\title{hpybdy: A set of functions for wishing someone happy birthday, nerd style}
\description{
The hpybdy package provides a function and set of base functions for
wishing someone special a happy birthday.
}
\section{Functions}{

determine_letter_index() - takes a character vector of length 1 and determines
the index number of each letter in the character in the letters object.

wish_happy_bday() - generates a plot that wishes someone happy bday in festive
colors.
}

