% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wish_happy_bdy.R
\name{wish_happy_bday}
\alias{wish_happy_bday}
\title{Wish happy birthday.}
\usage{
wish_happy_bday(name)
}
\arguments{
\item{name}{A character vector of length 1}
}
\description{
Generates a plot wishing a specific individual happy birthday.
}
\section{Warning}{

This function does not yet support names with spaces
}

